# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Live Review Foundation",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/llm/settings": {
      "get": {
        "tags": [
          "llm"
        ],
        "summary": "Read Settings",
        "operationId": "read_settings_api_v1_llm_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Public"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "llm"
        ],
        "summary": "Save Settings",
        "operationId": "save_settings_api_v1_llm_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Update"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Public"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "llm"
        ],
        "summary": "Clear Settings",
        "operationId": "clear_settings_api_v1_llm_settings_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Revision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Public"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/llm/settings/check": {
      "post": {
        "tags": [
          "llm"
        ],
        "summary": "Check Settings",
        "operationId": "check_settings_api_v1_llm_settings_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Check"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Public"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Check": {
        "properties": {
          "expected_revision": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Expected Revision"
          },
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "request_id"
        ],
        "title": "Check"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Public": {
        "properties": {
          "revision": {
            "type": "string",
            "title": "Revision",
            "default": "0"
          },
          "base_url": {
            "type": "string",
            "title": "Base Url",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "timeout_seconds": {
            "type": "integer",
            "title": "Timeout Seconds",
            "default": 60
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "status": {
            "type": "string",
            "enum": [
              "not_configured",
              "unverified",
              "checking",
              "verified",
              "check_failed",
              "unknown"
            ],
            "title": "Status",
            "default": "not_configured"
          },
          "checked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checked At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Usage"
              },
              {
                "type": "null"
              }
            ]
          },
          "debug_http": {
            "type": "boolean",
            "title": "Debug Http",
            "default": false
          },
          "analysis_enabled": {
            "type": "boolean",
            "const": false,
            "title": "Analysis Enabled",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Public"
      },
...
      "Revision": {
        "properties": {
          "expected_revision": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Expected Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision"
        ],
        "title": "Revision"
      },
...
      "Update": {
        "properties": {
          "expected_revision": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Expected Revision"
          },
          "base_url": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Base Url"
          },
          "model": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Model"
          },
          "timeout_seconds": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Timeout Seconds"
          },
          "api_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_revision",
          "base_url",
          "model",
          "timeout_seconds"
        ],
        "title": "Update"
      },
...
      "Usage": {
        "properties": {
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Usage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}